#92370 Subtitles are shown on the normal time (with 0.25 marge)

+++ src/Three/helper.wall.garbage.ts
@@ -3,6 +3,7 @@ import TaggingService, { Tags } from '@/services/TaggingService';
 import ThreeService from '@/services/ThreeService';
 import AnimationDefaults from './defaults.animation';
 import Development from './defaults.development';
+import TaggingHelper from './helper.tagging';
 
 export type GarabageHelperForWall = {
   removeGroupsByTag: (_tag: Tags) => void;
@@ -51,9 +52,11 @@ const WallGarbageHelper = (threeService: ThreeService, taggingService: TaggingSe
   };
 
   const newStorySelected = async () => {
-    const groupOfAssetsTags = taggingService.getByTag(Tags.GroupOfAssets);
-    await CustomAnimation().fadeOutGroups([groupOfAssetsTags[0].object], 0, AnimationDefaults.values.fadeStep);
-    threeService.RemoveFromScene(groupOfAssetsTags[0].object);
+    const groupOfAssetsTags = TaggingHelper(taggingService).getByTag(Tags.GroupOfAssets);
+    if (groupOfAssetsTags) {
+      await CustomAnimation().fadeOutGroups([groupOfAssetsTags[0].object], 0, AnimationDefaults.values.fadeStep);
+      threeService.RemoveFromScene(groupOfAssetsTags[0].object);
+    }
     storyCircle();
     removeArrayOfGroupsByTag(Tags.PauseScreenCenterText)
     removeByTag(Tags.PauseScreenBanner);
